fix logHelper and gwt permission when loging

+++ POS.Common/Helper/AuditLog/AuditLogMain.cs
@@ -8,7 +8,7 @@ namespace POS.Common.Helper.AuditLog
 {
     public class AuditLogMain
     {
-        public long AuditLogMainID { get; set; }
+        public int AuditLogMainID { get; set; }
         public string InputObject { get; set; }
 
         public int UserId { get; set; } = 0;
