+++ POS.Common/Helper/AuditLog/ExceptionLog.cs
@@ -8,7 +8,7 @@ namespace POS.Common.Helper.AuditLog
 {
     public class ExceptionLog
     {
-        public long ExceptionLogID { get; set; }
+        public int ExceptionLogID { get; set; }
         public string InputObject { get; set; }
         public string ExceptionMessage { get; set; }
         public int UserId { get; set; } = 0;
